# Improve performance of re-used symmetric one shots