@vcsjones 2023-10-06
Okay, going to close this for now, as the threading part needs a bit of work. Will re-open when that's in better shape, with new benchmarks.